Move store to src folder
