+++ src/main.js
@@ -3,7 +3,7 @@ import App from '@/App.vue';
 import SelectBoxIt from '@/components/common/SelectBoxIt.vue';
 import Vuelidate from 'vuelidate';
 import router from '@/router';
-import store from '@/store/store';
+import store from '@/store';
 import '@/registerServiceWorker';
 import apolloProvider from '@/apollo';
 import i18n from '@/i18n/i18n';
